==> 04_Calculate_pollen_percentage_unbinned.R <==
# This script filters the pollen data, applies the ppe correction and calculates 
# pollen percentages, unbinned, for use in the GAMS
# Annegreet Veeken, last changed 5-8-2021

# Libraries
library(tidyverse)
library(readxl)
library(xlsx)

lPOL <- readRDS("RDS_files/03_PollenWEU-Harmonised.rds")
ppe <- read_xlsx("Data/RPP_Dataset_v1_Table_3_4.xlsx")
ppe <- tibble(ppe.taxon = ppe$`Target taxon`,
              PPE = ppe$`RPP v1 (Northern Hemisphere)`)
taxa <- lPOL %>% purrr::map(., ~colnames(.)) %>% unlist() %>% unique()
harm <- read_xlsx("Data/HarmonizationTablePollen.xlsx")
harm <- harm %>% 
  filter(AccVarName2 %in% taxa) %>% 
  dplyr::select(AccVarName2, GroupId, ppe.taxon) %>% 
  distinct() 

# convert data frames to long format
lPOLlong <- lPOL %>% 
  purrr::map(.,
             ~dplyr::select_if(., !names(.) %in%
                                 c(".id", "age.old", "age.young",
                                   "date.type", "lat", "long", "dataset"))) %>%
  purrr::map(., ~gather(., key = "taxa", value = "count", -c(1:4)))

# add ppe taxon and group id 
lPOLlong <- lPOLlong %>% 
  purrr::map(. %>% 
               # WATCH OUT FOR DUPLICATION (check length of each df in list before and after join)
               left_join(., harm, by = c("taxa" = "AccVarName2")) %>% 
               # filter out non-pollen
               filter(., !ppe.taxon %in%  "not applicable") %>% 
               # filter out ages past 10000
               filter(., age <= 10000)) 

lPOLlong <- lPOLlong %>% 
  # filter no ppe data
  purrr::map(., ~ filter(., !ppe.taxon %in%  c("no ppe")) %>% 
               # select relevant columns
               dplyr::select(site.name, age, depth, time.bin, 
                             taxa, GroupId, ppe.taxon,
                             count))

# calculate percentage ppe corrected
lPOLcorrected <-  lPOLlong %>%
  purrr::map(. %>%
               # bind ppe data
               left_join(., ppe, by = "ppe.taxon") %>% 
               # adjust count with ppe
               mutate(adjustedcount = count*PPE) %>%
               group_by(age) %>% 
               # calculate percentages
               mutate(adjustedpercent = adjustedcount/sum(adjustedcount),
                      percent = count/sum(count, na.rm = TRUE)) %>%
               group_by(age, ppe.taxon) %>%
               summarise(adjustedpercent = sum(adjustedpercent, na.rm = TRUE),
                         percent = sum(percent)) %>% 
               # sort
               arrange(age, ppe.taxon))

# check if adjusted.percent > 0, otherwise PPE is probably missing
# x <- lPOLcorrected %>% purrr::map(., ~group_by(., age) %>%
#                                     summarise(., sum = sum(.$adjustedpercent)) %>%
#                                     filter(sum == 0))

saveRDS(lPOLcorrected, "RDS_files/04_PollenWEU-PPEcorrected-unbinned.rds")


lPOLcorrected_herbs <-  lPOLlong %>%
  purrr::map(. %>%
               # bind ppe data
               left_join(., ppe, by = "ppe.taxon") %>% 
               # filter for herbs 
               filter(GroupId == "HERB") %>% 
               # adjust count with ppe
               mutate(adjustedcount = count*PPE) %>%
               group_by(age) %>% 
               # calculate percentages
               mutate(adjustedpercent = adjustedcount/sum(adjustedcount),
                      percent = count/sum(count, na.rm = TRUE)) %>%
               group_by(age, ppe.taxon) %>%
               summarise(adjustedpercent = sum(adjustedpercent, na.rm = TRUE),
                         percent = sum(percent)) %>% 
               # sort
               arrange(age, ppe.taxon))

saveRDS(lPOLcorrected_herbs, "RDS_files/04_PollenWEU-PPEcorrected-HERBS.rds")

lPOLcorrected_trees <-  lPOLlong %>%
  purrr::map(. %>%
               # bind ppe data
               left_join(., ppe, by = "ppe.taxon") %>% 
               # filter for trees 
               filter(GroupId == "TRSH") %>% 
               # adjust count with ppe
               mutate(adjustedcount = count*PPE) %>%
               group_by(age) %>% 
               # calculate percentages
               mutate(adjustedpercent = adjustedcount/sum(adjustedcount),
                      percent = count/sum(count, na.rm = TRUE)) %>%
               group_by(age, ppe.taxon) %>%
               summarise(adjustedpercent = sum(adjustedpercent, na.rm = TRUE),
                         percent = sum(percent)) %>% 
               # sort
               arrange(age, ppe.taxon))

saveRDS(lPOLcorrected_trees, "RDS_files/04_PollenWEU-PPEcorrected-TRSH.rds")
